unify kmeans and metis construction
